requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
openai-whisper==20231117
torch==2.1.2
python-multipart==0.0.6
python-dotenv==1.0.0

# Para análise de vídeo
ffmpeg-python==0.2.0
openai==1.10.0
